test/requirements_init.txt
annotated-types==0.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.4
colorama==0.4.6
Flask==2.2.5
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
imageio==2.21.3
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.1.2
Levenshtein==0.20.9
MarkupSafe==2.1.3
numpy==1.25.0
opencv-python==4.5.5.64
packaging==23.1
pandas==1.4.4
Pillow==9.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.8.2
python-Levenshtein==0.20.9
pytz==2023.3
pyzbar==0.1.9
rapidfuzz==2.15.1
requests==2.32.1
rsa==4.9
six==1.16.0
tqdm==4.66.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==2.3.6
